Refactor music/tone playing to a separate class.
