+++ YtaRobot/src/main/cpp/YtaRobot.cpp
@@ -477,16 +477,17 @@ void YtaRobot::TeleopPeriodic()
     ShootSequence();
     //HoodSequence();
     //HangSequence();
+    CheckForManualAdjust();
+    UpdateSmartDashboard();
 
     //PneumaticSequence();
     
     CameraSequence();
 
+    // These only do things if their configs are enabled.
+    // See YtaLed.hpp and YtaMusic.hpp for the controls.
     LedSequence();
-    //MusicSequence();
-
-    CheckForManualAdjust();
-    UpdateSmartDashboard();
+    MusicSequence();
 }
 
 
@@ -928,79 +929,28 @@ void YtaRobot::LedSequence()
 ////////////////////////////////////////////////////////////////
 void YtaRobot::MusicSequence()
 {
-    // Note: The control mode for the motors can only be one
-    //       thing at a time.  Using a motor for acutal motion
-    //       will not work at the same time as playing tones.
-    static bool bMusicPlaying = false;
-    if (m_pDriveController->GetButtonState(PLAY_MUSIC_BUTTON))
+    if (Yta::Music::Config::PLAYING_TONES_ENABLED)
     {
-        bMusicPlaying = true;
-    }
-
-    // Not playing any music, just return
-    if (!bMusicPlaying)
-    {
-        return;
-    }
-
-    // Add more of these as needed
-    static const MusicTone noNote(units::frequency::hertz_t(0));
-    static const MusicTone cNote(units::frequency::hertz_t(262));
-    static const MusicTone dNote(units::frequency::hertz_t(294));
-    static const MusicTone eNote(units::frequency::hertz_t(330));
-    static const MusicTone fNote(units::frequency::hertz_t(349));
-    static const MusicTone gNote(units::frequency::hertz_t(392));
-    static const MusicTone aNote(units::frequency::hertz_t(440));
-    static const MusicTone bNote(units::frequency::hertz_t(494));
-    static const MusicTone CNote(units::frequency::hertz_t(523));
-
-    static Timer * pMusicTimer = new Timer();
-    static units::time::second_t lastNotePlayTime = 0.0_s;
-    static bool bInit = false;
-    static uint32_t noteIndex = 0U;
-
-    if (!bInit)
-    {
-        pMusicTimer->Start();
-        bInit = true;
-    }
-
-    struct NoteControl
-    {
-        MusicTone m_Tone;
-        units::time::second_t m_LengthSeconds;
-    };
-
-    // Change the notes and the lengths in here to make a song.
-    // You can add or remove or change as many you want.
-    // If you need more tones (frequencies in hertz), add them up above.
-    // If you want a pause, use noNote and a length.
-    static NoteControl scaleNotes[] =
-    {
-        {noNote, 100_ms},
-        {cNote, 100_ms},
-        {dNote, 100_ms},
-        {eNote, 100_ms},
-        {fNote, 100_ms},
-        {gNote, 100_ms},
-        {aNote, 100_ms},
-        {bNote, 100_ms},
-        {CNote, 100_ms},
-        {noNote, 100_ms},
-    };
-    static const size_t MAX_NOTE_INDEX = sizeof(scaleNotes) / sizeof (NoteControl);
-
-    if ((pMusicTimer->Get() - lastNotePlayTime) > scaleNotes[noteIndex].m_LengthSeconds)
-    {
-        noteIndex++;
-        if (noteIndex >= MAX_NOTE_INDEX)
+        // Note: The control mode for the motors can only be one
+        //       thing at a time.  Using a motor for acutal motion
+        //       will not work at the same time as playing tones.
+        static bool bPlayMusic = false;
+        if (m_pDriveController->DetectButtonChange(PLAY_MUSIC_BUTTON))
         {
-            noteIndex = 0U;
-            bMusicPlaying = false;
+            bPlayMusic = true;
         }
-        // Pick a motor to play a sound
-        //m_pMusicMotor->SetControl(scaleNotes[noteIndex].m_Tone);
-        lastNotePlayTime = pMusicTimer->Get();
+
+        if (bPlayMusic)
+        {
+            // Note: Change the nullptr to the TalonFX object for the
+            //       motor to play tones on!  This code will crash otherwise.
+
+            // When PlayTones() returns false, the music is over
+            if (!YtaMusicController::PlayTones(nullptr))
+            {
+                bPlayMusic = false;
+            }
+        }
     }
 }
